requirements.txt
pyyaml
requests
argparse
pylint
typing
